+++ docs/changes.rst
@@ -6,6 +6,7 @@ Not yet released.
 * Added support for sharing a component between projects.
 * Fixed multiple unnamed variables check behavior with multiple format flags.
 * Dropped mailing list field on project in favor of generic instructions for translators.
+* Added pseudolocale generation addon.
 
 Weblate 4.4.2
 -------------